Stage 2/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
SQLAlchemy==2.0.44
pydantic==2.12.3
pydantic-settings==2.1.1
requests==2.32.5
pillow==10.4.0
colorlog==6.10.1
mysql-connector-python==9.5.0
fastapi-limiter==0.1.6
redis==5.0.7
pytest==8.4.2
pytest-asyncio==1.2.0
